+++ domain/src/main/java/me/androidbox/models/CurrentModel.kt
@@ -1,4 +1,4 @@
 package me.androidbox.models
 
-data class CurrentModel(val temperatureInCelsius: Float)
+data class CurrentModel(val temperatureInCelsius: Int)
